+++ include/endstone/endstone.hpp
@@ -103,6 +103,8 @@ static_assert(_ITERATOR_DEBUG_LEVEL == 0,
 #include "event/event_handler.h"
 #include "event/event_priority.h"
 #include "event/handler_list.h"
+#include "event/inventory/inventory_event.h"
+#include "event/inventory/inventory_interact_event.h"
 #include "event/level/dimension_event.h"
 #include "event/level/level_event.h"
 #include "event/player/player_arm_swing_event.h"